feeback improvements, history now works with screens, fix for buttons not working after screens were added
git-svn-id: file:///home/svn/omero/trunk@4781 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/tools/OmeroImporter/src/ome/formats/importer/HistoryDB.java
@@ -345,8 +345,8 @@ public class HistoryDB implements IObservable
         return rs.getInt(1);
     }
     
-    public int insertFileHistory(long importID, long experimenterID, int rowNum, 
-            String filename, long projectID, long datasetID, String status, File file) 
+    public int insertFileHistory(Integer importID, Long experimenterID, Integer rowNum, 
+            String filename, Long projectID, Long datasetID, String status, File file) 
     throws SQLException
     {
         return update(
